requirements.txt
Django==5.1
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
